refactor(tests): remove redundant test modules and consolidate backend and type tests into dedicated files for improved organization and maintainability
